bitkeeper revision 1.1153 (41126318o43x3YiPCk8UwYtIQ4XdtA)
Fixes to the warping mechanism in BVT

+++ tools/python/xen/xend/server/SrvDomain.py
@@ -112,8 +112,8 @@ class SrvDomain(SrvDir):
                      ['mcuadv', 'int'],
                      ['warpback', 'int'],
                      ['warpvalue', 'int'],
-                     ['warpl', 'int'],
-                     ['warpu', 'int']])
+                     ['warpl', 'long'],
+                     ['warpu', 'long']])
         val = fn(req.args, {'dom': self.dom.name})
         return val
